requirements.txt
chromadb>=0.4.0
sentence-transformers>=2.2.2
langchain>=0.1.0
langchain-community>=0.0.20
ollama>=0.1.7
pyttsx3>=2.90
SpeechRecognition>=3.10.0
pyaudio>=0.2.13
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyMuPDF>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-docx>=0.8.11
pdf2image>=1.16.3
numpy>=1.24.0
tqdm>=4.66.0
gTTS>=2.3.2
langdetect>=1.0.9
pygame>=2.5.0
feedparser>=6.0.10
requests>=2.31.0
APScheduler>=3.10.4
Flask>=3.0.0
Werkzeug>=3.0.0
pandas>=2.0.0
openpyxl>=3.1.0
pytest>=7.4.0
# vosk>=0.3.45  # Optional: uncomment for full offline STT support
#               # Requires model download: https://alphacephei.com/vosk/models
#               # Recommended: vosk-model-small-en-in-0.4 + vosk-model-small-ta-0.4
